==> nb2_GBDT_and_preprocessing_increment.ipynb <==
# Build plain Random Forest classifier on Tf-Idf

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import confusion_matrix,classification_report

# https://www.kaggle.com/code/neerajmohan/nlp-text-classification-using-tf-idf-features

t = pd.read_excel(f'{sdir}t4m_2022_05_05__21_59_v3.xlsx')

df = t[['text', 'class_name']].copy()
df['label'] = df['class_name'].map(cls2idx)
df['label'].value_counts()

tmp = df[['text','label']].copy()

data2 = tmp.copy()


data2 = data2[['text', 'label']].set_axis(['headline', 'label'], axis=1)
data2=Dataset.from_pandas(data2)

# 80% train, 20% test + validation
train_testvalid = data2.train_test_split(test_size=0.3,seed=15)

# Split the 10% test + valid in half test, half valid
test_valid = train_testvalid['test'].train_test_split(test_size=0.5,seed=15)

# gather everything to have a single DatasetDict
data2 = DatasetDict({
    'train': train_testvalid['train'],
    'test': test_valid['test'],
    'valid': test_valid['train'],
                   })

tfidf_vectorizer = TfidfVectorizer() 

Xy_train = train_testvalid['train'].to_pandas()
Xy_test = test_valid['test'].to_pandas()
Xy_val = test_valid['train'].to_pandas()

X_train, y_train = Xy_train['headline'], Xy_train['label']
X_val, y_val = Xy_val['headline'], Xy_val['label']
X_test, y_test = Xy_test['headline'], Xy_test['label']

X_train = pd.concat([X_train, X_val], ignore_index=True)
y_train = y_train.values.tolist() + y_val.values.tolist()

tfidf_train_vectors = tfidf_vectorizer.fit_transform(X_train)

tfidf_test_vectors = tfidf_vectorizer.transform(X_test)

classifier = RandomForestClassifier()

classifier.fit(tfidf_train_vectors,y_train)

y_pred = classifier.predict(tfidf_test_vectors)
print(classification_report(y_test,y_pred, target_names=[v.replace('_', ' ') for i,v in idx2cls.items()]))

# Build CatBoostClassifier on Tf-Idf

from catboost import CatBoostClassifier
from sklearn.utils.class_weight import compute_class_weight

tfidf_vectorizer = TfidfVectorizer() 

Xy_train = train_testvalid['train'].to_pandas()

###############

Xy = Xy_train.copy()
target = 'headline'

cnts = pd.DataFrame(Xy['label'].value_counts()).reset_index().set_axis(['cname', 'cnt'], axis=1).\
    assign(needed=lambda row: row['cnt'].max()-row['cnt'])

rows2add = []
for ind, row in cnts.iterrows():
    smpl = Xy.query(f'label == {row["cname"]}').sample(row['needed'], replace=True)
    for ind2, row2 in smpl.iterrows():
        rows2add.append([random_insert2str(row2[target]), row2['label']])

Xy = pd.concat([Xy, pd.DataFrame(rows2add, columns = Xy.columns)]).sample(frac=1)

Xy_train = Xy.copy()

###############


Xy_test = test_valid['test'].to_pandas()
Xy_val = test_valid['train'].to_pandas()

X_train, y_train = Xy_train['headline'], Xy_train['label']
X_val, y_val = Xy_val['headline'], Xy_val['label']
X_test, y_test = Xy_test['headline'], Xy_test['label']

tfidf_train_vectors = tfidf_vectorizer.fit_transform(X_train)
tfidf_test_vectors = tfidf_vectorizer.transform(X_test)
tfidf_val_vectors = tfidf_vectorizer.transform(X_val)

classes = np.unique(y_train)
weights = compute_class_weight(class_weight='balanced', classes=classes, y=y_train)
class_weights = dict(zip(classes, weights))

params_cat = {
      'eval_metric':'TotalF1', 
      'iterations':5000,
      'verbose':500,
      'early_stopping_rounds':550,
      'border_count': 254,
      'use_best_model':True,
      'class_weights':class_weights,
    'task_type':'CPU'
  }


model = CatBoostClassifier(**params_cat)
preds = model.fit(tfidf_train_vectors, y_train,
                 eval_set=(tfidf_val_vectors, y_val)
                 ).predict(tfidf_test_vectors)

print(classification_report(y_test, preds, target_names=[v.replace('_', ' ') for i,v in idx2cls.items()]))

# Prepare for random insert

rand_ins = [
    'app store',
    'google play',
    'burger king',
#     'macdonalds',
#     'rostics',
#     'kfc',
    'ресторан стоит',
    'телефон работает',
    'еду кушают',
    'ночью спят',
    'днем работают',
    'обед днем',
    'завтрак утром',
    'ужин вечером'
]


def random_insert2str(s0):
    
    s = s0.strip()
    spaces = [i for i, ltr in enumerate(s) if ltr == ' ']
    inserts = []
    if len(spaces) > 0:
        inserts=np.random.choice(rand_ins, len(spaces)//5 + 1)
    else:
        return s + ' ' + np.random.choice(rand_ins, 1)[0]

    spaces2ins = np.random.choice(spaces, len(inserts))

    ss = ''
    i0 = 0
    for ind, i in enumerate(spaces2ins):
        ss += s[i0:i] + f' {inserts[ind]}'
        i0 = i
    if i0 < len(s):
        ss += s[i0:]
    return ss

# Load file and define function to compare F1 with different data

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import confusion_matrix,classification_report
from datasets import load_dataset,Dataset,DatasetDict, load_metric
import numpy as np
import pandas as pd

# https://www.kaggle.com/code/neerajmohan/nlp-text-classification-using-tf-idf-features

t = pd.read_excel(f'{sdir}t4m_2022_05_05__21_59_v3.xlsx')
cnames = list(set(t['class_name'].values))
idx2cls = {i:v for i,v in enumerate(cnames)}
cls2idx = {v:i for i,v in enumerate(cnames)}

df = t[['content', 'text_1', 'text_2', 'text_3_0', 'text_3', 'text_5', 'text_6', 'class_name']].copy()
df['label'] = df['class_name'].map(cls2idx)
df['label'].value_counts()

tmp = df[['content', 'text_1', 'text_2', 'text_3_0', 'text_3', 'text_5', 'text_6','label']].copy()


data2 = tmp.copy()


data2 = data2[['content', 'text_1', 'text_2', 'text_3_0', 'text_3', 'text_5', 'text_6', 'label']].copy()
data2=Dataset.from_pandas(data2)

# 80% train, 20% test + validation
train_testvalid = data2.train_test_split(test_size=0.3,seed=15)

# Split the 10% test + valid in half test, half valid
test_valid = train_testvalid['test'].train_test_split(test_size=0.5,seed=15)

# gather everything to have a single DatasetDict
data2 = DatasetDict({
    'train': train_testvalid['train'],
    'test': test_valid['test'],
    'valid': test_valid['train'],
                   })

# tfidf_vectorizer = TfidfVectorizer() 

Xy_train = train_testvalid['train'].to_pandas()
Xy_test = test_valid['test'].to_pandas()
Xy_val = test_valid['train'].to_pandas()


def score_text(target, out=False, random_insert=False):
    X_train, y_train = Xy_train[target], Xy_train['label']
    X_val, y_val = Xy_val[target], Xy_val['label']
    X_test, y_test = Xy_test[target], Xy_test['label']

    X_train = pd.concat([X_train, X_val], ignore_index=True)
    y_train = y_train.values.tolist() + y_val.values.tolist()
    

    ##################################

    if random_insert:
        Xy = pd.DataFrame({target:X_train.values.tolist(), 'label':y_train})

        cnts = pd.DataFrame(Xy['label'].value_counts()).reset_index().set_axis(['cname', 'cnt'], axis=1).\
            assign(needed=lambda row: row['cnt'].max()-row['cnt'])

        rows2add = []
        for ind, row in cnts.iterrows():
            smpl = Xy.query(f'label == {row["cname"]}').sample(row['needed'], replace=True)
            for ind2, row2 in smpl.iterrows():
                rows2add.append([random_insert2str(row2[target]), row2['label']])

        Xy = pd.concat([Xy, pd.DataFrame(rows2add, columns = Xy.columns)]).sample(frac=1)

        X_train = Xy[target]
        y_train = Xy['label'].values.tolist()

    ##################################

    tfidf_vectorizer = TfidfVectorizer() 
    
    tfidf_train_vectors = tfidf_vectorizer.fit_transform(X_train)

    tfidf_test_vectors = tfidf_vectorizer.transform(X_test)

    classifier = RandomForestClassifier()

    classifier.fit(tfidf_train_vectors,y_train)

    y_pred = classifier.predict(tfidf_test_vectors)
    if out: print(classification_report(y_test,y_pred))
    
    return classification_report(y_test,y_pred, output_dict=True, target_names=list(cls2idx.keys()))

def score_text2(target, n=15, random_insert=False):
    res = []
    for i in range(n):
        res.append(score_text(target, random_insert=random_insert)['macro avg']['f1-score'])
        
    return np.mean(np.array(res)), np.std(np.array(res))

# Find out F1 increment for each data preparation step

for param in ['content', 'text_1', 'text_2', 'text_3_0', 'text_3', 'text_5', 'text_6']:
    f,s = score_text2(param, n=25)
    print(f'testing {param}: F1 = {f:.3f} ± {s:.3f}')

# Random insert increment

param = 'text_5'
f,s = score_text2(param, n=5, random_insert=True)
print(f'testing {param}: F1 = {f:.3f} ± {s:.3f}')

# Random insert increment wo business replacements

param = 'text_6'
f,s = score_text2(param, n=5, random_insert=True)
print(f'testing {param}: F1 = {f:.3f} ± {s:.3f}')
