==> nb6_multilabel_scoring.ipynb <==
%load_ext autoreload
%autoreload 2

import numpy as np
import pandas as pd

sdir = './scrub/'

# read data, see value counts per class

data = pd.read_csv(f'multilabel_v1.csv')[['content', 'target']].sample(frac=1)
data['target'] = data['target'].apply(lambda x: eval(x))

t = pd.DataFrame(data['target'].values.tolist(), columns = [f'cls_{i+1}' for i in range(15)])
t.sum()

# load init dataset to form class dicts

t = pd.read_excel(f'{sdir}t4m_2022_05_05__21_59.xlsx')
cnames = list(set(t['class_name'].values))
idx2cls = {i:v for i,v in enumerate(cnames)}
cls2idx = {v:i for i,v in enumerate(cnames)}
display(idx2cls)
cls2idx

from datasets import load_dataset,Dataset,DatasetDict, load_metric

from transformers import DataCollatorWithPadding,AutoModelForSequenceClassification, Trainer, TrainingArguments,AutoTokenizer,AutoModel,AutoConfig
from transformers.modeling_outputs import TokenClassifierOutput
import torch
import torch.nn as nn


# form dataset in transformer style

data = data[['content', 'target']].set_axis(['headline', 'label'], axis=1)
data=Dataset.from_pandas(data)

# 80% train, 20% test + validation
train_testvalid = data.train_test_split(test_size=0.3,seed=15)

# Split the 10% test + valid in half test, half valid
test_valid = train_testvalid['test'].train_test_split(test_size=0.5,seed=15)

# gather everything to have a single DatasetDict
data = DatasetDict({
    'train': train_testvalid['train'],
    'test': test_valid['test'],
    'valid': test_valid['train'],
                   })

data['train'][0]

# next section is similar to nb4, with custom bert. We build model, obtain predictions, print classification report

BATCH_SIZE = 8
seq_len = 512
n_classes = len(idx2cls.items())
out_features=768

# device = torch.device('mps')
# couldn't make distilbert uncased work on apple silicon, so i took random bert trained on news.
# also found bert trained on rotten tomatoes, but it was big and slow
checkpoint = 'cointegrated/rubert-tiny2'
# checkpoint = 'cointegrated/rubert-tiny'
checkpoint = 'DeepPavlov/distilrubert-small-cased-conversational'
tokenizer = AutoTokenizer.from_pretrained(checkpoint)
tokenizer.model_max_len=seq_len


def tokenize(batch):
    return tokenizer(batch["headline"], truncation=True,max_length=seq_len)

tokenized_dataset = data.map(tokenize, batched=True)

tokenized_dataset.set_format("torch",columns=["input_ids", "attention_mask", "label"])
data_collator = DataCollatorWithPadding(tokenizer=tokenizer)


class CustomModel(nn.Module):
    def __init__(self, checkpoint, num_labels, out_features=768):
        super(CustomModel,self).__init__()
        self.num_labels = num_labels

        #Load Model with given checkpoint and extract its body
        self.model = AutoModel.from_pretrained(checkpoint,
                                                config=AutoConfig.from_pretrained(checkpoint, 
                                                                                 output_attentions=True,
                                                                                 output_hidden_states=True))
        
        self.out_features = out_features#model.encoder.layer[1].output.dense.out_features
        self.dropout = nn.Dropout(0.1)
        self.classifier = nn.Linear(self.out_features, num_labels) # load and initialize weights

    def forward(self, input_ids=None, attention_mask=None,labels=None):
        #Extract outputs from the body
        outputs = self.model(input_ids=input_ids, attention_mask=attention_mask)

        #Add custom layers
        sequence_output = self.dropout(outputs[0]) #outputs[0]=last hidden state
        
        logits = self.classifier(sequence_output[:,0,:].view(-1,self.out_features)) # calculate losses

        loss = None
        if labels is not None:
            loss_fct = nn.CrossEntropyLoss()

            loss = loss_fct(logits.view(-1, self.num_labels), labels.float().view(-1, self.num_labels))

        return TokenClassifierOutput(loss=loss, logits=logits, hidden_states=outputs.hidden_states,attentions=outputs.attentions)

device = torch.device("cuda" if torch.cuda.is_available() else "mps")
model = CustomModel(checkpoint=checkpoint, num_labels=n_classes, out_features=out_features).to(device)

from torch.utils.data import DataLoader

train_dataloader = DataLoader(
    tokenized_dataset["train"], shuffle=True, batch_size=BATCH_SIZE, collate_fn=data_collator
)
eval_dataloader = DataLoader(
    tokenized_dataset["valid"], batch_size=BATCH_SIZE, collate_fn=data_collator
)

from transformers import AdamW, get_scheduler

optimizer = AdamW(model.parameters(), lr=5e-5)

num_epochs = 3
num_training_steps = num_epochs * len(train_dataloader)
lr_scheduler = get_scheduler(
    "linear",
    optimizer=optimizer,
    num_warmup_steps=0,
    num_training_steps=num_training_steps,
)

from tqdm.auto import tqdm

metric = load_metric("accuracy")

progress_bar_train = tqdm(range(num_training_steps))
progress_bar_eval = tqdm(range(num_epochs * len(eval_dataloader)))


for epoch in range(num_epochs):
    model.train()
    for batch in train_dataloader:
        batch = {k: v.to(device) for k, v in batch.items()}
        outputs = model(**batch)
        loss = outputs.loss
        loss.backward()

        optimizer.step()
        lr_scheduler.step()
        optimizer.zero_grad()
        progress_bar_train.update(1)

    model.eval()
    for batch in eval_dataloader:
        batch = {k: v.to(device) for k, v in batch.items()}
        with torch.no_grad():
            outputs = model(**batch)

        logits = outputs.logits
        predictions = torch.argmax(logits, dim=-1)
        targets = torch.argmax(batch["labels"], dim=-1)
        metric.add_batch(predictions=predictions, references=targets)
        progress_bar_eval.update(1)

    print(metric.compute())


torch.save(model, 'bert_model_v1.pt')

metric = load_metric("accuracy")

model.eval()

test_dataloader = DataLoader(
    tokenized_dataset["test"], batch_size=BATCH_SIZE, collate_fn=data_collator
)

for batch in test_dataloader:
    batch = {k: v.to(device) for k, v in batch.items()}
    with torch.no_grad():
        outputs = model(**batch)

    logits = outputs.logits

    predictions = torch.argmax(logits, dim=-1)
    targets = torch.argmax(batch["labels"], dim=-1)
    metric.add_batch(predictions=predictions, references=targets)

print(metric.compute())

# F1

metric = load_metric("recall")

model.eval()

test_dataloader = DataLoader(
    tokenized_dataset["test"], batch_size=BATCH_SIZE, collate_fn=data_collator
)

preds = []
labs = []
for batch in test_dataloader:
    batch = {k: v.to(device) for k, v in batch.items()}
    with torch.no_grad():
        outputs = model(**batch)

    logits = outputs.logits

    predictions = torch.argmax(logits, dim=-1)
    targets = torch.argmax(batch["labels"], dim=-1)
    metric.add_batch(predictions=predictions, references=targets)
    
    preds += list(torch.argmax(logits, dim=-1).detach().cpu().numpy())
    labs += list(torch.argmax(batch["labels"], dim=-1).detach().cpu().numpy())

print(metric.compute(average='macro'))

test = tokenized_dataset["test"].to_pandas()
test['preds'] = preds
test['true'] = labs

from sklearn.metrics import classification_report

print(classification_report(test['true'], test['preds']))

# Next we need to calculate positional F1 metric, and compare with 95% in multi-class approach

# metric = load_metric("recall")

model.eval()

test_dataloader = DataLoader(
    tokenized_dataset["test"], batch_size=BATCH_SIZE, collate_fn=data_collator
)

preds = []
labs = []
thresh = 0.85
for batch in test_dataloader:
    batch = {k: v.to(device) for k, v in batch.items()}
    with torch.no_grad():
        outputs = model(**batch)

    logits = outputs.logits
    l = torch.sigmoid(logits).cpu().numpy().tolist()
    
    predictions = [[1 if i > thresh else 0 for i in j] for j in l]
#     break

#     predictions = torch.argmax(logits, dim=-1)
#     targets = torch.argmax(batch["labels"], dim=-1)
    targets = batch["labels"].cpu().numpy().tolist()
    preds.extend(predictions)
    labs.extend(targets)
    
    
#     preds += list(torch.argmax(logits, dim=-1).detach().cpu().numpy())
#     labs += list(torch.argmax(batch["labels"], dim=-1).detach().cpu().numpy())


test = tokenized_dataset["test"].to_pandas()
test['preds'] = preds
test['true'] = labs

from sklearn.metrics import classification_report

# print(classification_report(test['true'], test['preds']))

test.head()

true0 = test['true'].values[0]
pred0 = test['preds'].values[0]

def pos_recall(true0, pred0):
    r = 0
    for i, v in enumerate(true0):
        if pred0[i] == v : r += 1
        else: r -+ 1
    return r/len(true0)

res = []
for true0, pred0 in zip(test['true'].values, test['preds'].values):
    res.append(pos_recall(true0, pred0))
np.mean(np.array(res))

# comparisson is in favour of multi-class
